% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{aggregate_cells}
\alias{aggregate_cells}
\title{Aggregate cells}
\usage{
aggregate_cells(
  .data,
  .sample = NULL,
  slot = "data",
  assays = NULL,
  aggregation_function = Matrix::rowSums,
  .metadata_columns_to_keep = NULL
)
}
\arguments{
\item{.data}{A tidySingleCellExperiment object}

\item{.sample}{A vector of variables by which cells are aggregated}

\item{slot}{The slot to which the function is applied}

\item{assays}{The assay to which the function is applied}

\item{aggregation_function}{The method of cell-feature value aggregation}

\item{.metadata_columns_to_keep}{a column vector e.g. c(column_1, column_2), which combinqtion is unique for `.sample` to be included in the pseudobulk metadata. If NULL those columns will be automatically detected, which operation can be slow for large objects.}
}
\value{
A tibble object
}
\description{
Combine cells into groups based on shared variables and aggregate feature counts.
}
\examples{
print("this is a method generics Example is not applicable")
# <object> |> aggregate_cells(c(groups, ident), assays = "counts")

}
