% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{join_features}
\alias{join_features}
\title{join_features}
\usage{
join_features(
  .data,
  features = NULL,
  all = FALSE,
  exclude_zeros = FALSE,
  shape = "long",
  ...
)
}
\arguments{
\item{.data}{A tidy SingleCellExperiment object}

\item{features}{A vector of feature identifiers to join}

\item{all}{If TRUE return all}

\item{exclude_zeros}{If TRUE exclude zero values}

\item{shape}{Format of the returned table "long" or "wide"}

\item{...}{Parameters to pass to join wide, i.e. assay name to extract feature abundance from and gene prefix, for shape="wide"}
}
\value{
A `tbl` containing the information.for the specified features
}
\description{
join_features() extracts and joins information for specific
  features
}
\details{
This function extracts information for specified features and
  returns the information in either long or wide format.
}
\examples{

print("this is a method generics Example is not applicable")
# <object> |> join_features(features=c("HLA-DRA", "LYZ"))

}
