% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_all_related_columns}
\alias{select_all_related_columns}
\title{Subset columns}
\usage{
select_all_related_columns(.data, .column)
}
\arguments{
\item{.data}{A tibble}

\item{.column}{A vector of column names}
}
\value{
A tibble
}
\description{
Subset columns
}
